Yay, finally I have a test backend that is not as hard to maintain as the aRts one. Using the old test programs and the new fake backend I found that moc doesn't fully expand macros as I assumed when I wrote phonondefs.h. Well you shouldn't assume and not do a simple test. Anyway, I moved the slot definitions out of the macros and into a common base class.
Also there was another problem with the this pointer being used before
properly being accessible in the constructor. (Obvious when you think about
it.) So I changed the whole construction and deleting process of the iface
objects once again. This time it looks maintainable and it works (well, at
least with the fake backend).

Also added a BUGS file to keep track of the bugs I know and don't want to fix
yet.

svn path=/branches/work/kdelibs-phonon/phonon/; revision=499861